BACKPORT: sched/eevdf: Fix avg_vruntime()
The expectation is that placing a task at avg_vruntime() makes it
eligible. Turns out there is a corner case where this is not the case.

Specifically, avg_vruntime() relies on the fact that integer division
is a flooring function (eg. it discards the remainder). By this
property the value returned is slightly left of the true average.

However! when the average is a negative (relative to min_vruntime) the
effect is flipped and it becomes a ceil, with the result that the
returned value is just right of the average and thus not eligible.

BUG=b:308209790
TEST=Manual tested on DUT

Fixes: af4cf40470c2 ("sched/fair: Add cfs_rq::avg_vruntime")
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Link: https://lore.kernel.org/all/169632975777.3135.1926466016577986434.tip-bot2@tip-bot2/
Signed-off-by: Youssef Esmat <youssefesmat@chromium.org>

Change-Id: If04b3946bd4f58a82de1c9673537a8d8655d53c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/5089903
Reviewed-by: Vineeth Pillai <vineethrp@google.com>
Commit-Queue: Youssef Esmat <youssefesmat@chromium.org>
Tested-by: Vineeth Pillai <vineethrp@google.com>